Refactor in to separate module. Convenience functions.
